requirements.txt
flask
flask-cors
requests
numpy
faiss-cpu
sentence-transformers
pickle5  # Only required for older Python versions, otherwise built-in
